fix(nlb/proto): effective_port int64->int32 (JSON parity with TargetGroup.port)
HealthCheck.effective_port was int64 -> proto3-JSON serialized it as a quoted
string "8443" while sibling TargetGroup.port (int32) serialized as a bare number
8443 -> two port fields, inconsistent JSON types. int32 is correct by-domain
(0..65535). Regen pkg/api stub + int32 cast at the sole type2pb assignment site.

buf-breaking FIELD_SAME_TYPE (int64->int32) — intra-redesign, pre-GA wire, no
external clients; effective_port is net-new vs main so zero breaking vs main.
RED-lock: services/nlb/tests/newman/cases/target-group.py:961/984 (expect number)
flips GREEN on nlb+gateway deploy. proto-api-reviewer APPROVED.

Follow-up (tracked): TcpOptions/HttpOptions/HttpsOptions/GrpcOptions.port still
int64 -> same JSON-number parity worth a uniform pass across the LB surface.

+++ services/nlb/internal/dto/type2pb/health_check.go
@@ -33,7 +33,7 @@ func healthCheckToPb(hc domain.HealthCheck, tgPort domain.LbPort) *lbv1.HealthCh
 		UnhealthyThreshold: int64(hc.UnhealthyThreshold),
 		HealthyThreshold:   int64(hc.HealthyThreshold),
 		// effective_port — output-only derived: probe-override || TG.port.
-		EffectivePort: int64(hc.EffectivePort(tgPort)),
+		EffectivePort: int32(hc.EffectivePort(tgPort)),
 	}
 	switch {
 	case hc.TCP != nil:
